Moved delete button to card details with a cancel button added
