+++ instaGreet/src/components/CardDetails.jsx
@@ -1,12 +1,13 @@
 import axios from "axios"
 import Card from "./Card"
-import { useParams } from 'react-router-dom'
+import { useParams, useNavigate } from 'react-router-dom'
 import { useState, useEffect } from "react"
 import NavBar from "./NavBar"
 
-const CardDetails = ({ token }) => {
+const CardDetails = ({ token, cardID, setCardID }) => {
     const [card, setCard] = useState([])
     const { id } = useParams()
+    const navigate = useNavigate()
 
     useEffect(() => {
         axios
@@ -17,15 +18,22 @@ const CardDetails = ({ token }) => {
             })
       }, [id])
 
+    const goToDelete = (event, card) => {
+        const key = card.id
+        setCardID(key)
+        navigate(`/delete-card/${key}`)
+    }
+
     return (
         <>
         <NavBar />
             <Card 
-            key={card.id}
-            front_text={card.front_text}
-            background_color={card.background_color}
-            creator={card.creator}
+                key={card.id}
+                front_text={card.front_text}
+                background_color={card.background_color}
+                creator={card.creator}
             />
+            <button onClick={((e) => goToDelete(e, card))}>Delete</button>
         </>
     )
 }
